requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
jinja2
python-multipart
markdown
openai
